hbase_data_extraction_service/requirements.txt
Django==3.1.3
djangorestframework==3.12.2
django-cors-headers==3.5.0
happybase
easybase
hbase-python
google
protobuf
kazoo
git+git://github.com/CurleySamuel/PyBase.git